% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinkr-eyer.R
\name{convert_to_eyer}
\alias{convert_to_eyer}
\title{Converts loaded data to EyerObject}
\usage{
convert_to_eyer(obj)
}
\arguments{
\item{obj}{list loaded by \code{\link{load_asc}} and preprocessed}
}
\value{
EyerObject
}
\description{
The EyerObject is a default eyetracking object to be used with the `eyer` package.
It is recommended that you preprocess your data before converting it to eyer, as it might be more
difficult to translate some data, especially events
}
\details{
the data is a list with fixations, gaze and events fields. It converts them to the
Eyer object structure and does little preprocessing
}
