% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_plot.R
\name{df_plot}
\alias{df_plot}
\title{Create a dataframe that is ready for ggplot}
\usage{
df_plot(eye_var, col_var, out_var, n)
}
\arguments{
\item{eye_var, col_var, out_var, n}{df with the indiidual data, name of the column of interest, name of output column, number of cases}
}
\value{
A df
}
\description{
Create a dataframe that is ready for ggplot
}
