% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fits.R
\name{get_fits}
\alias{get_fits}
\title{Get individual fits (linear and log)}
\usage{
get_fits(free, fixed)
}
\arguments{
\item{free, fixed}{Two dataframes: of the data grouped at the individual level -- one for Free, one for Fixed}
}
\value{
A dataframe with fits of each individual
}
\description{
Get individual fits (linear and log)
}
