% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slopes.R
\name{get_slopes}
\alias{get_slopes}
\title{Get individual slopes (linear and log)}
\usage{
get_slopes(free, fixed)
}
\arguments{
\item{free, fixed}{Two dataframes: of the data grouped at the individual level -- one for Free, one for Fixed}
}
\value{
A dataframe with slopes of each individual
}
\description{
Get individual slopes (linear and log)
}
