% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzmean.R
\name{nzmean}
\alias{nzmean}
\title{Calculate mean without zeros}
\usage{
nzmean(nums)
}
\arguments{
\item{nums}{Path to the input file}
}
\value{
A numeric
}
\description{
Takes a numeric vector (???) and first extracts all the zeros
If the vector consists of all zeros, return zero
If not, calculate the mean by excluding the zeros
Returns a numeric value
}
