% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr1000-loading.R
\name{parse_calibrations}
\alias{parse_calibrations}
\title{Reads in calibration}
\usage{
parse_calibrations(text)
}
\arguments{
\item{text}{character vector of the asc file}
}
\description{
Reads in calibration
}
