% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr1000-loading.R
\name{parse_events}
\alias{parse_events}
\title{Decodes events from passed sr1000 text file}
\usage{
parse_events(text)
}
\arguments{
\item{text}{}
}
\value{
data.frame
}
\description{
Used in case you load the text elsewhere or you just
want to decode parts of it. In case you want to parse entire file, you can
use read_events instead
}
