% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr1000-loading.R
\name{parse_fixations}
\alias{parse_fixations}
\title{Decodes fixations from the given character vector}
\usage{
parse_fixations(text)
}
\arguments{
\item{text}{}
}
\description{
Used in case you load the text elsewhere or you just
want to decode parts of it. In case you want to parse entire file, you can
use read_fixations instead
}
