% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr1000-loading.R
\name{parse_gaze}
\alias{parse_gaze}
\title{Decodes gaze information from passed parsed asc text}
\usage{
parse_gaze(text)
}
\arguments{
\item{text}{character vector of the asc file}
}
\value{
data.frame with the gaze
}
\description{
Used in case you load the text elsewhere or you just
want to decode parts of it. In case you want to parse entire file, you can
use read_gaze instead
}
