% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinkr-loading.R
\name{parse_logging_rate}
\alias{parse_logging_rate}
\title{Fetches logging frequency}
\usage{
parse_logging_rate(text)
}
\arguments{
\item{text}{loaded asc file as a text}
}
\value{
numeric rate of logging
}
\description{
Used in case you load the text elsewhere or you just
want to decode the frequency. Goes line by line through parts of the text.
IN the asc, the frequnecy is found in line "SAMPLES	GAZE	RIGHT	RATE	1000.00 ...."
}
\details{
information from the passed asc text
}
