% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr1000-loading.R
\name{read_events}
\alias{read_events}
\title{Reads in events from the given filepath}
\usage{
read_events(filepath)
}
\arguments{
\item{filepath}{path to the asc file}
}
\value{
data.frame with the events
}
\description{
Reads in events from the given filepath
}
