% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr1000-loading.R
\name{read_resolution}
\alias{read_resolution}
\title{match two digits at least three digit long after Display coords
? signifies non greedy match (as least as possible)}
\usage{
read_resolution(filepath)
}
\description{
match two digits at least three digit long after Display coords
? signifies non greedy match (as least as possible)
}
