% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinkr-preprocessing.R
\name{remove_event_brackets}
\alias{remove_event_brackets}
\title{Removes brackets from the parsed events}
\usage{
remove_event_brackets(df)
}
\arguments{
\item{df}{event data.frame as loaded by `read_events` or `parse_events`}
}
\description{
Removes brackets from the parsed events
}
