% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinkr-preprocessing.R
\name{remove_key_up}
\alias{remove_key_up}
\title{Helper to remove all depression of keys}
\usage{
remove_key_up(df)
}
\arguments{
\item{df}{event data.frame as loaded by `read_events` or `parse_events`}
}
\description{
Eyelink events log both key down and key_up events.
these are for many purposes unnecessary, so we can half the number
of logged events by removing key up (user releases key) events
}
