% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr1000-loading.R
\name{load_asc_file}
\alias{load_asc_file}
\title{Loads in sr 1000 asc file and outputs a list}
\usage{
load_asc_file(filepath)
}
\arguments{
\item{filepath}{path to the asc file}
}
\description{
Loads in sr 1000 asc file and outputs a list
}
