% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr1000-loading.R
\name{read_fixations}
\alias{read_fixations}
\title{Reads eye fixations from a asc file}
\usage{
read_fixations(filepath)
}
\arguments{
\item{filepath}{Path to the asc file}
}
\description{
Reads eye fixations from a asc file
}
