% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinkr-preprocessing.R
\name{remove_event_keys}
\alias{remove_event_keys}
\title{Removes events with unwanted keys}
\usage{
remove_event_keys(df, keys)
}
\arguments{
\item{keys}{character vector of all the keys you want to remove. e.g c("w", "a"). If you didn't
use the remove bracket function, these need to be in the [w] format}

\item{df_events}{df event data.frame as loaded by `read_events` or `parse_events`}
}
\description{
Removes events with unwanted keys
}
